chore: move ava testing setup into separate rollup config
